requirements.txt
Django==5.1.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.3.0
django-filter==24.1
markdown==3.5.2
dj-rest-auth==5.0.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
django-cleanup==8.1.0 
weasyprint